Moved ui templates into seperate components folder
