+++ nodes/modules/system/domain/models/template/components/index.js
@@ -1,5 +1,5 @@
-
+const ui = require('./ui');
 
 module.exports = {
-
+  ui
 };
