+++ nodes/modules/system/domain/models/template/components/ui/index.js
@@ -1,0 +1,5 @@
+const motion = require('./motion');
+
+module.exports = {
+  motion
+};
