+++ nodes/modules/system/domain/models/template/components/ui/motion/index.js
@@ -1,0 +1,5 @@
+const lighting = require('./lighting');
+
+module.exports = {
+  lighting
+};
