+++ nodes/modules/system/domain/models/template/components/ui/motion/lighting/index.js
