+++ nodes/modules/system/domain/models/template/dynamic/area/climate/index.js
