+++ nodes/modules/system/domain/models/template/dynamic/area/index.js
