+++ nodes/modules/system/domain/models/template/dynamic/area/lighting/index.js
