+++ nodes/modules/system/domain/models/template/dynamic/area/motion/detection/index.js
