+++ nodes/modules/system/domain/models/template/dynamic/area/motion/index.js
