+++ nodes/modules/system/domain/models/template/dynamic/area/motion/lighting/MotionLightingTargetStateAreaInputText.js
