+++ nodes/modules/system/domain/models/template/dynamic/area/motion/lighting/index.js
@@ -1,4 +1,3 @@
-const ui = require('./ui');
 const MotionLightingHybridTargetAreaInputText = require("./MotionLightingHybridTargetAreaInputText");
 const MotionLightingHybridTargetAreaTemplateSelect = require("./MotionLightingHybridTargetAreaTemplateSelect");
 const MotionLightingHybridTargetStateAreaInputText = require("./MotionLightingHybridTargetStateAreaInputText");
@@ -9,7 +8,6 @@ const MotionLightingTargetStateAreaInputText = require("./MotionLightingTargetSt
 const MotionLightingTimeoutAreaInputNumber = require("./MotionLightingTimeoutAreaInputNumber");
 
 module.exports = {
-    ui,
     MotionLightingHybridTargetAreaInputText,
     MotionLightingHybridTargetAreaTemplateSelect,
     MotionLightingHybridTargetStateAreaInputText,
