+++ nodes/modules/system/domain/models/template/dynamic/index.js
@@ -1,5 +1,7 @@
+const area = require("./area");
 //const AverageMetricSensor = require("./AverageMetricSensor");
 
 module.exports = {
+    area
     //AverageMetricSensor,
 };
