+++ nodes/modules/system/domain/models/template/index.js
@@ -1,11 +1,9 @@
-const area = require("./area");
 const components = require("./components");
 const dynamic = require("./dynamic");
 //const Template = require("./Template");
 const TemplateGenerator = require("./TemplateGenerator");
 
 module.exports = {
-    area,
     components,
     dynamic,
     //Template,
