+++ nodes/modules/system/test/unit/domain/models/template/area/motion/lighting/MotionLightingHybridTargetAreaInputText.test.js
@@ -2,7 +2,7 @@ const fs = require('fs');
 const path = require('path');
 const { mockAreas, mockEntities, normalizeMultilineString } = require('../../../../../../../../util/test');
 
-const MotionLightingHybridTargetAreaInputText = require('../../../../../../../../domain/models/template/area/motion/lighting/MotionLightingHybridTargetAreaInputText');
+const MotionLightingHybridTargetAreaInputText = require('../../../../../../../../domain/models/template/dynamic/area/motion/lighting/MotionLightingHybridTargetAreaInputText');
 
 describe('MotionLightingHybridTargetAreaInputText', () => {
