+++ nodes/modules/system/test/unit/domain/models/template/area/motion/lighting/MotionlightingTimeoutAreaInputDatetime.test.js
@@ -2,7 +2,7 @@ const fs = require('fs');
 const path = require('path');
 const { mockAreas, mockEntities, normalizeMultilineString } = require('../../../../../../../../util/test');
 
-const MotionLightingTimeoutAreaInputNumber = require('../../../../../../../../domain/models/template/area/motion/lighting/MotionLightingTimeoutAreaInputNumber');
+const MotionLightingTimeoutAreaInputNumber = require('../../../../../../../../domain/models/template/dynamic/area/motion/lighting/MotionLightingTimeoutAreaInputNumber');
 
 describe('MotionLightingTimeoutAreaInputNumber', () => {
